projector pass stub
git-svn-id: http://svn.sgscript.org/tacstrike@150 89f9ca44-2ff2-4b64-a49d-2a4f1987ee72

+++ tools/blendcubedds.cpp
@@ -87,7 +87,7 @@ int main( int argc, char* argv[] )
 		return 1;
 	}
 	
-	if( WIDTH * WIDTH * 4 != sz / 6 )
+	if( size_t(WIDTH * WIDTH * 4) != sz / 6 )
 	{
 		fprintf( stderr, "unsupported input format\n" );
 		return 1;
